backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
fonttools>=4.55.0
brotli>=1.1.0
aiofiles>=24.1.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
